Rename VersatileThermostat with BaseThermostat. Tests ok
